requirements.txt
# Production dependencies
kopf==1.38.0
kubernetes==34.1.0
prometheus_client==0.23.1
PyYAML==6.0.3
boto3==1.40.64
botocore==1.40.64
werkzeug==3.1.3
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0

